RDKB-6628: Log if SSID are same or different for 2.4GHz and 5GHz
Reason for change:
Change is done to be able to collect stat from field devices to undersatnd if user has configured different SSIDs for 2.4Ghz and 5GHz.
Log message will be periodically(every 6 hours) printed to indicate whether SSIDs are same or different.
Test Procedure:
1) Put the script into target and anullay triggered the script rxtx100.sh
2) Verified prints are logged into ArmConsolelog.txtx.0 file as below
When both SSIDs are same:
2016-07-12:12:44:37:520000 [RDKB_STAT_LOG] 2.4G and 5G SSIDs are same

When both SSIDs are different:
2016-07-12:12:45:18:640000 [RDKB_STAT_LOG] 2.4G and 5G SSIDs are different

When SSIDs are not fetched from WiFi agent:
2016-07-12:12:49:13:700000 [RDKB_PLATFORM_ERROR] Didn't get WiFi 2.4 GHz SSID from agent
2016-07-12:12:49:13:850000 [RDKB_PLATFORM_ERROR] Didn't get WiFi 5 GHz SSID from agent

Change-Id: If4ae31aa002ed2fa2df603fac464bdb8537a044a
Risks: None
Signed-off-by: achelakkode<anoopnair_c@infosys.com>