RDKB-9204: SNMP reset takes more than 1 minute
Reason for change: When reboot is triggered through SNMP, it takes close to 1 minute for reboot.
Out of this, RDKB logging scripts take anywhere between 25-30 seconds. This is because of DCA synchronization, log backup etc.
To be able to reduce this reboot time, sysdescriptor value population is pushed during boot up and tarring of log files is also done during bootup.

Test Procedure:
1) Send reboot through SNMP
2) Measure reboot time from the time SNMP command reached the unit
3) Get the time stamps from console and verify the time spend in RDKB scripts.

With the fix, it is verified that RDK logger scripts are now taking 15 seconds to execute logging activities.

Change-Id: I5cdcbc48765bc14ede089767661ea583ea775931
Risks: None
Signed-off-by: achelakkode<anoopnair_c@infosys.com>