RDKB-9391: Managing tar balls when DCM settings changes
Reason for change: When DCM settings are received, based on the regular upload flag we need to manage the tar balls to be uploaded.
As per the implementation, when DCM configuration is set to "false" for log upload, then it removes all tar files from nvram2 including reboot tar files.
We should not be removing reboot tar files from nvram2.

Test Procedure:
1) When DCM settings is "true" , verify if log upload after 1.5MB is success.
2) If DCM settings for log upload is "false", verified log upload was success only for reboot
and tar ball itself was not created in case of 1.5MB

Change-Id: I30e9b25c154789c99c5b7444d7f48520cd4742bb
Risks: None
Signed-off-by: achelakkode<anoopnair_c@infosys.com>