RDKB-27900:systemctl: not found in ArmConsole Logs
Reason for change: systemd is not running on ARM so systemctl command is not present in arm side, When we are executing  script in arm side since the command is not present we are getting the error message on console.

Test Procedure: 1. enable snmpv3 on box
                2. execute script(handlesnmpv3.sh), with the fix we are not getting the error message on console.
Signed-off-by: sohail_ahmad@cable.comcast.com
Risks: Low

(cherry picked from commit 3fba1715f271a76462b30eddf62e58716c6e52c4)
Change-Id: Ic2b9010e4cc7f97ca440c8d5473870be60603ae8