RDKBDEV-641: read from /version.txt instead of /fss/gw/version.txt
Access the actual version.txt file directly rather than relying on
the /fss/gw/version.txt symlink.

more robust parsing of imagename from /version.txt

The eventual goal is to be able to source /version.txt to extract the
image name (instead of needing to parse it with grep, sed, etc). As a
transitional step towards that, update parsing to support all
combinations of : or = as separator and string quoted or not quoted.
ie /version.txt containing any of:

 1: imagename:mng-image-rdkb-puma6-arm-qemuarm-20210326015040
 2: imagename=mng-image-rdkb-puma6-arm-qemuarm-20210326015040
 3: imagename:"mng-image-rdkb-puma6-arm-qemuarm-20210326015040"
 4: imagename="mng-image-rdkb-puma6-arm-qemuarm-20210326015040"

Currently /version.txt contains 1), but will eventually become 4).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I69af8f3fbd2e03e86024c7768380b86b7d52540b